add authorization checks for reply crud operations, add 'isPublished' helper method in Thread model, add new permissions related to reply

+++ app/Http/Controllers/ReplyController.php
@@ -15,7 +15,15 @@ class ReplyController extends Controller
 {
     public function index(int $threadId): JsonResponse
     {
-        $thread = Thread::with('replies.user')->findOrFail($threadId);
+        $user = Auth::guard('sanctum')->user();
+        $thread = null;
+        $relations = ['replies.user'];
+
+        if($user?->can('view all threads')) {
+            $thread = Thread::with($relations)->findOrFail($threadId);
+        } else {
+            $thread = Thread::with($relations)->whereNotNull('published_at')->findOrFail($threadId);
+        }
 
         return new JsonResponse([
             'replies' => ReplyResource::collection($thread->replies)
@@ -24,9 +32,12 @@ class ReplyController extends Controller
 
     public function store(ReplyStoreRequest $request, int $threadId): JsonResponse
     {
+        $thread = Thread::findOrFail($threadId);
+
+        $this->authorize('storeReply', $thread);
+
         $reply = new Reply($request->validated());
         $reply->user()->associate(Auth::user());
-        $thread = Thread::findOrFail($threadId);
         $thread->replies()->save($reply);
         $reply->save();
 
@@ -39,6 +50,9 @@ class ReplyController extends Controller
     public function update(ReplyUpdateRequest $requet, int $id): JsonResponse
     {
         $reply = Reply::findOrFail($id);
+
+        $this->authorize('update', $reply);
+
         $reply->update($requet->validated());
         $reply->save();
 
@@ -51,6 +65,9 @@ class ReplyController extends Controller
     public function destroy(int $id): JsonResponse
     {
         $reply = Reply::findOrFail($id);
+
+        $this->authorize('delete', $reply);
+
         $reply->delete();
 
         return new JsonResponse([
